requirements.txt
Flask==3.1.0
qrcode[pil]
pillow==11.1.0
pyzbar==0.1.9
cryptography==44.0.0